+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Add embedded-hal 1 implementations for the following drivers:
 Introduce LPSPI improvements:
 
 - Add additional checks for LPSPI frame sizes.
-- Rework LPSPI clock settings.
+- Rework LPSPI clock settings, initialization, and disable.
 - Add `flush()` method.
 
 ## [0.5.4] 2023-11-26